requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
python-multipart==0.0.9
slowapi==0.1.9
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.12
gpt4all==2.5.1
PyPDF2==3.0.1
sentence-transformers==3.1.1
faiss-cpu==1.8.0
numpy==1.26.4
streamlit==1.39.0
requests==2.32.3
